pyproject.toml
[tool.poetry]
name = "atk-training-rin-q-basic"
version = "0.2.0"
description = ""
authors = ["Harine <haariinee@gmail.com>"]
readme = "README.md"
packages = [{include = "rin_db_exc"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
typer = "^0.9.0"
uvicorn = "^0.23.1"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
atk-training-rin-q-basic = "rin_db_exc.main_cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
